# api/shareloop.openapi.yaml
openapi: "3.0.3"
info:
  title: Shareloop API
  version: "1.0.0"
  description: API for the Shareloop application
servers:
  - url: /api
paths:
  /featured-items:
    post:
      summary: Get featured items
      description: Returns a list of featured items
      operationId: getFeaturedItems
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LatLng"
      responses:
        "200":
          description: A list of featured items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemOverview"
        "500":
          description: Internal server error
  /items/search:
    post:
      summary: Search items
      description: Search items with filters
      operationId: searchItems
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemSearchRequest"
      responses:
        "200":
          description: A list of matching items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemOverview"
        "500":
          description: Internal server error
  /items:
    post:
      summary: Create a new item
      description: Creates a new item listing
      operationId: createItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
      responses:
        "201":
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateItemResponse"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
  /items/{itemId}:
    get:
      summary: Get item details
      description: Returns full details for a single item
      operationId: getItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemDetail"
        "404":
          description: Item not found
        "500":
          description: Internal server error
    put:
      summary: Update an item
      operationId: updateItem
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemRequest"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateItemResponse"
        "401":
          description: Unauthorized
        "403":
          description: Not the owner
        "404":
          description: Item not found
        "500":
          description: Internal server error
...
  /items/{itemId}/edit:
    get:
      summary: Get item edit details (owner only)
      description: Returns full item details including location for the item owner
      operationId: getItemEdit
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item edit details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemEditDetail"
        "401":
          description: Unauthorized
        "403":
          description: Not the owner
        "404":
          description: Item not found
        "500":
          description: Internal server error
  /items/{itemId}/images:
    post:
      summary: Upload an image for an item
      operationId: uploadItemImage
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadItemImageRequest"
      responses:
        "201":
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadItemImageResponse"
        "401":
          description: Unauthorized
        "403":
          description: Not the owner
        "404":
          description: Item not found
        "500":
          description: Internal server error
    put:
      summary: Edit item images (reorder / delete)
      operationId: editItemImages
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditItemImagesRequest"
      responses:
        "204":
          description: Images updated
        "401":
          description: Unauthorized
        "403":
          description: Not the owner
        "404":
          description: Item not found
        "500":
          description: Internal server error
...
  /auth/refresh:
    post:
      summary: Refresh tokens
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid or expired refresh token
        "500":
          description: Internal server error
...
components:
  schemas:
...
    ItemOverview:
      type: object
      required: [id, title, description, author, score, category]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        author:
          $ref: "#/components/schemas/Person"
        distance:
          $ref: "#/components/schemas/Distance"
          nullable: true
        city:
          type: string
        postalCode:
          type: string
        address:
          type: string
          nullable: true
        score:
          type: number
          format: double
        imageUuid:
          type: string
          nullable: true
        category:
          type: string
        pricePerDay:
          type: number
          format: double
    Person:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    Distance:
      type: object
      required: [km]
      properties:
        km:
          type: number
          format: double
    LatLng:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
...
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    User:
      type: object
      required: [id, email, lastOnlineAt, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        lastOnlineAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    LoginResult:
      type: object
      required: [user, accessToken, refreshToken, accessExpiresAt, refreshExpiresAt]
      properties:
        user:
          $ref: "#/components/schemas/User"
        accessToken:
          type: string
        refreshToken:
          type: string
        accessExpiresAt:
          type: string
          format: date-time
        refreshExpiresAt:
          type: string
          format: date-time
    CreateItemRequest:
      type: object
      required: [title, description, city, postalCode, lat, lng, category]
      properties:
        title:
          type: string
        description:
          type: string
        city:
          type: string
        postalCode:
          type: string
        address:
          type: string
          nullable: true
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        category:
          type: string
        pricePerDay:
          type: number
          format: double
    CreateItemResponse:
      type: object
      required: [id]
      properties:
        id:
          type: integer
    ItemDetail:
      type: object
      required: [id, title, description, author, score, createdAt, imageUuids, category, itemRatingCount, itemRatings]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        author:
          $ref: "#/components/schemas/Person"
        score:
          type: number
          format: double
        city:
          type: string
        postalCode:
          type: string
        address:
          type: string
          nullable: true
        imageUuids:
          type: array
          items:
            type: string
            format: uuid
        category:
          type: string
        pricePerDay:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        itemRatingCount:
          type: integer
          minimum: 0
        itemRatings:
          type: array
          items:
            $ref: "#/components/schemas/ItemRating"
    ItemEditDetail:
      type: object
      required: [id, title, description, author, score, createdAt, imageUuids, lat, lng, category]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        author:
          $ref: "#/components/schemas/Person"
        score:
          type: number
          format: double
        city:
          type: string
        postalCode:
          type: string
        address:
          type: string
          nullable: true
        imageUuids:
          type: array
          items:
            type: string
            format: uuid
        category:
          type: string
        pricePerDay:
          type: number
          format: double
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
    UploadItemImageRequest:
      type: object
      required: [data, filename, sortOrder]
      properties:
        data:
          type: string
          description: Base64-encoded image data
        filename:
          type: string
          description: Original filename for extension detection
        sortOrder:
          type: integer
          description: Position of this image in the item's image list
    UploadItemImageResponse:
      type: object
      required: [uuid]
      properties:
        uuid:
          type: string
          format: uuid
    UpdateItemRequest:
      type: object
      required: [title, description, city, postalCode, lat, lng, category]
      properties:
        title:
          type: string
        description:
          type: string
        city:
          type: string
        postalCode:
          type: string
        address:
          type: string
          nullable: true
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        category:
          type: string
        pricePerDay:
          type: number
          format: double
    ItemSearchRequest:
      type: object
      properties:
        query:
          type: string
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        maxDistanceKm:
          type: number
          format: double
        categories:
          type: array
          items:
            type: string
        minScore:
          type: number
          format: double
        sortBy:
          type: string
          enum: [relevance, distance, score, newest]
    ReorderEntry:
      type: object
      required: [uuid, sortOrder]
      properties:
        uuid:
          type: string
          format: uuid
        sortOrder:
          type: integer
    EditItemImagesRequest:
      type: object
      required: [reorder, delete]
      properties:
        reorder:
          type: array
          items:
            $ref: "#/components/schemas/ReorderEntry"
        delete:
          type: array
          items:
            type: string
            format: uuid
...
    ItemRating:
      type: object
      required: [id, rentRequestId, itemId, reviewer, condition, cleanliness, overall, createdAt]
      properties:
        id:
          type: integer
        rentRequestId:
          type: integer
        itemId:
          type: integer
        reviewer:
          $ref: "#/components/schemas/Person"
        condition:
          type: integer
          minimum: 1
          maximum: 5
        cleanliness:
          type: integer
          minimum: 1
          maximum: 5
        overall:
          type: number
          format: double
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token